// src/FirstTracks.Web/Scripts/Models/SkiResortSnowfall/SkiResortSnowfall.d.ts
export declare class SkiResortSnowfall {
    skiResortSnowfallId: string;
    month: number;
    year: number;
    inches: number;
    createdOn: Date;
    modifiedOn: Date;
    constructor(skiResortSnowfallId: string, month: number, year: number, inches: number, createdOn: Date, modifiedOn: Date);
}
export default SkiResortSnowfall;